notify user api
